peptide_atlas/pyproject.toml
[project]
name = "peptide-atlas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "biopython>=1.85",
    "pandas>=2.3.0",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]
